Fix: restore the InsertOutcome re-export and split the context module
The rebase over the harness-context wrapper changes dropped
`pub use insert_outcome::InsertOutcome;` from `context/mod.rs` while leaving
the `mod insert_outcome;` declaration and `lib.rs`'s re-export in place, so the
crate no longer compiled: `unresolved import context::InsertOutcome`. This only
surfaced in the Windows coverage leg because that leg is the first to build the
crate from scratch in CI.

Restore the re-export next to the module declaration. With the crate compiling
again, Whitaker could finally measure the module and reported `context` at 414
lines, over the 400-line limit, so move the `FixtureRef`/`FixtureRefMut` borrow
guards into `context/fixture_ref.rs` and re-export them. Their public paths are
unchanged.
